+++ src/main/java/data_access/movies/MovieDBDataAccessObject.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.movies;
 
 import java.io.IOException;
 import java.util.ArrayList;
